iter12: use transaction and insert+conflict in JSONUpdateMetricsHandler
